Use Base64 URL decoder for reading keys
* Improve testability of the Key resolver
* Improve logging on validation failures
