+++ prototype/northbound/security/openid-connect/integration-test.bndrun
@@ -26,6 +26,8 @@
 	junit-platform-commons;version='[1.9.0,1.9.1)',\
 	junit-platform-engine;version='[1.9.0,1.9.1)',\
 	junit-platform-launcher;version='[1.9.0,1.9.1)',\
+	net.bytebuddy.byte-buddy;version='[1.12.13,1.12.14)',\
+	net.bytebuddy.byte-buddy-agent;version='[1.12.13,1.12.14)',\
 	org.apache.felix.configadmin;version='[1.9.24,1.9.25)',\
 	org.apache.felix.scr;version='[2.2.2,2.2.3)',\
 	org.eclipse.jetty.alpn.client;version='[11.0.13,11.0.14)',\
@@ -39,6 +41,9 @@
 	org.eclipse.sensinact.gateway.northbound.security.authentication-api;version='[0.0.2,0.0.3)',\
 	org.eclipse.sensinact.gateway.northbound.security.openid-connect;version='[0.0.2,0.0.3)',\
 	org.eclipse.sensinact.gateway.northbound.security.openid-connect-tests;version='[0.0.2,0.0.3)',\
+	org.mockito.junit-jupiter;version='[4.7.0,4.7.1)',\
+	org.mockito.mockito-core;version='[4.7.0,4.7.1)',\
+	org.objenesis;version='[3.2.0,3.2.1)',\
 	org.opentest4j;version='[1.2.0,1.2.1)',\
 	org.osgi.service.component;version='[1.5.0,1.5.1)',\
 	org.osgi.test.common;version='[1.2.1,1.2.2)',\
