+++ prototype/northbound/security/openid-connect/src/test/resources/public-key-examples/example-key.json
@@ -1,0 +1,17 @@
+{
+  "keys": [
+    {
+      "kid": "FwTNsWCd4YNEIS491YEi2eyNkYheqpWkN9kWAXYHcMc",
+      "kty": "RSA",
+      "alg": "RS256",
+      "use": "sig",
+      "n": "g3Yy1E00o3QpHCxiKO1fN-hyH_U9OkB8wil8MOi2JbB5Vj2RLxCSmNmWGMXW9XTrm7Or03nBt789eFKmPK7zolG3E2HvCpuk-mAU1uSBMYaTt6CmdnQMoDDsotPHeNDT43TMw7ey2dfaionGPHx7YYAldPhiJ_sehcRrAuP1Ad33rkIMUmmbVkV8nhQkW736zqoo1OZdl4b7VDIypwycAk5qT5rS0rcW088RRroqLxqc8SGRN2PMLYy7l-3bp2u9sw_AihVVXRdkk_iN3j7ON8GWKcAbc6BcMb1m8J8pbTqikjNia6_e5vKUyYFRoIQuUlIrT4JaYJgOTGNgDp9l9w",
+      "e": "AQAB",
+      "x5c": [
+        "MIIClzCCAX8CBgGITU9PPTANBgkqhkiG9w0BAQsFADAPMQ0wCwYDVQQDDAR0ZXN0MB4XDTIzMDUyNDEwMjg0M1oXDTMzMDUyNDEwMzAyM1owDzENMAsGA1UEAwwEdGVzdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAIN2MtRNNKN0KRwsYijtXzfoch/1PTpAfMIpfDDotiWweVY9kS8QkpjZlhjF1vV065uzq9N5wbe/PXhSpjyu86JRtxNh7wqbpPpgFNbkgTGGk7egpnZ0DKAw7KLTx3jQ0+N0zMO3stnX2oqJxjx8e2GAJXT4Yif7HoXEawLj9QHd965CDFJpm1ZFfJ4UJFu9+s6qKNTmXZeG+1QyMqcMnAJOak+a0tK3FtPPEUa6Ki8anPEhkTdjzC2Mu5ft26drvbMPwIoVVV0XZJP4jd4+zjfBlinAG3OgXDG9ZvCfKW06opIzYmuv3ubylMmBUaCELlJSK0+CWmCYDkxjYA6fZfcCAwEAATANBgkqhkiG9w0BAQsFAAOCAQEASeBxa/rknqDLUsdSPTzTWWj6C4y8u2H6D/doLtw3c6/Fku3cEWZLuLdgWkc2uupBKwJTS2eY0ErXG6B8y2ZqnoHdFPTJeXy5cn5GUgmmmMLqvX33rdgQKrxMbwKWHvMsSRsRIGg8YVYwHRUq1GyhJ3VvX1YZAvfPnkwXXL8LatuPu1xMPrIdLGigHrZAbtqQDXqOyJlfUI+UoVM7MlbODkz17Rgl0Yi+ZWjrwu0ORxYLg3RA91PfHR+SEPC+M2wC+vD0aZxup4hjNbaORMBuICL8wSEn2hPW5tSHE7fD1eiGSQPF39yToU3Hfcaweh1guckQar/fdWvP4kMgCAgZoQ=="
+      ],
+      "x5t": "2LPSpkkuXQrObe8cDRnpZA7ZN_M",
+      "x5t#S256": "KuX8NxKBUj-s_xs1niOw43zoEQB3SpKNlHAYNXnfpWI"
+    }
+  ]
+}
