+++ prototype/northbound/security/openid-connect/src/test/resources/public-key-examples/example-key2.json
@@ -1,0 +1,30 @@
+{
+  "keys": [
+    {
+      "kid": "kZPqM0qrcmkhgAn_AheCvuHv_cxMmU966q3XU0-K1y8",
+      "kty": "RSA",
+      "alg": "RSA-OAEP",
+      "use": "enc",
+      "n": "icdf_Qa-XK5ITdYfABRqx0PCi3NxplZs5Phoa3lWniZRZi1TJb7641gZUB4o0wMxr8om20-L-gvEVQMuv1jGRs4Lh8FZ_OEzuQ5URnTo7JQzdPkiCYTvErLpCrBgzWr4nakyuR4yaYAEER5R_MiuFjTHoWCMj-q3hDWW1TBVxZBYfv_wDZ0e_CRRU3TO7kCmWVukSI-4VDuOi46FbihD_b0LPQeRLYIMx9eNlau_x25UhGCWpo5KC0N4u5pARHD5bCCqRleulZZa4X05bqUuV_1-tZ9SAoDQFY4saiDK-y3Ves8HTrY7bDxGefVq3qWCBf94Urme1GrCN-6_G8Xecw",
+      "e": "AQAB",
+      "x5c": [
+        "MIIClzCCAX8CBgGITknT+zANBgkqhkiG9w0BAQsFADAPMQ0wCwYDVQQDDAR0ZXN0MB4XDTIzMDUyNDE1MDIyMVoXDTMzMDUyNDE1MDQwMVowDzENMAsGA1UEAwwEdGVzdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAInHX/0GvlyuSE3WHwAUasdDwotzcaZWbOT4aGt5Vp4mUWYtUyW++uNYGVAeKNMDMa/KJttPi/oLxFUDLr9YxkbOC4fBWfzhM7kOVEZ06OyUM3T5IgmE7xKy6QqwYM1q+J2pMrkeMmmABBEeUfzIrhY0x6FgjI/qt4Q1ltUwVcWQWH7/8A2dHvwkUVN0zu5ApllbpEiPuFQ7jouOhW4oQ/29Cz0HkS2CDMfXjZWrv8duVIRglqaOSgtDeLuaQERw+WwgqkZXrpWWWuF9OW6lLlf9frWfUgKA0BWOLGogyvst1XrPB062O2w8Rnn1at6lggX/eFK5ntRqwjfuvxvF3nMCAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAKZKcNIsPN0vXq5+v6o2oOUt01XqQbCV2IZ36tDH5R6i7UvMZsgEy+sGmqSAs0wLuE+henpChZRPvP12Vmpo+Hye3tcMbjG4tY2qq5xTytoy+MRzZfnFR0JTci7Lx04KFelDiQ/PVPxjCgmuHJ9eRKTzJTFmNOEcdlo1CDe+h3zub7VHCR1giQpwHouItGOWqQqFpLX0laALrRsBhsbmEpB4dll+Ehdta8OBy1DYgAW3/MBeP34HQPEPOyj9DUEGSLsGFA+Oog1Y7nPtLva68oXrLK0OrcgQSdoM+EQE4htfdVPcZ2mJqH+jXeCSQrW7yofubv7QS3VdFVFsrntZfaA=="
+      ],
+      "x5t": "3T7raVrfOrGNgWdQZusLlsA-uJk",
+      "x5t#S256": "HGalBXGqE1cuRpl9-CClZhFMPBIQOB_lZWIOGJsT6lY"
+    },
+    {
+      "kid": "nC_0bveSKUkxshBTTzhjRo1wu-p7xmqc6FDHT9AiG1I",
+      "kty": "RSA",
+      "alg": "RS256",
+      "use": "sig",
+      "n": "rXvxtYdvCXzkHOiPrq84V02oAo5L6sGzelc3ApoN6Ke0e1Vof_7Jhb91qqZqt9ZY6oOEIq3onhY_vbcmL1ZhSwXGbIbpKx-R7Q_D4ymHqtCg1dCqpEE07UdNJDWWSBAuVsT53rjIGJImL2XsiElhZ9qzXgZQm5_UrPGtNsHDJ6hZ_t6o-IeDkIo5wxd4gWk9derS8qAyOHh_4nx6HHDRYZcxCqYxkHe-WyQAYlwGitZb1OfWsMFqIcgic95hxmXEClGWX9ATkiw_E7War7Qij1pduaxL55tq3i6cvIgZm1onH3TpRYW7cN54Pm-PQ_cnihFb3VnvRy0yAjGN4Xus-w",
+      "e": "AQAB",
+      "x5c": [
+        "MIIClzCCAX8CBgGITknSCjANBgkqhkiG9w0BAQsFADAPMQ0wCwYDVQQDDAR0ZXN0MB4XDTIzMDUyNDE1MDIyMFoXDTMzMDUyNDE1MDQwMFowDzENMAsGA1UEAwwEdGVzdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAK178bWHbwl85Bzoj66vOFdNqAKOS+rBs3pXNwKaDeintHtVaH/+yYW/daqmarfWWOqDhCKt6J4WP723Ji9WYUsFxmyG6Ssfke0Pw+Mph6rQoNXQqqRBNO1HTSQ1lkgQLlbE+d64yBiSJi9l7IhJYWfas14GUJuf1KzxrTbBwyeoWf7eqPiHg5CKOcMXeIFpPXXq0vKgMjh4f+J8ehxw0WGXMQqmMZB3vlskAGJcBorWW9Tn1rDBaiHIInPeYcZlxApRll/QE5IsPxO1mq+0Io9aXbmsS+ebat4unLyIGZtaJx906UWFu3DeeD5vj0P3J4oRW91Z70ctMgIxjeF7rPsCAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAGEQxR85EfaJXzBPLO7u3USnm+QMmM/jTbAgvWLjlwNadRxb2RVEakK5W4rgdQQzjmnUKZL+mgca5WA8IMBqZm7mWwq+NAfF/Hc0Y0tNHqQ1qD0mPCqua+uv6fxShFBeBcV1QlR9Be2jmmUBQn260HMrIIBLBj2RrlHssDcVGun9xCGjlmVkxslHNvRJJ651nxK1aa4MrFH0NYuAcYuElq+cmcW06EXfBPi6Xp7fJ7nk7kptKqALpb8zXJOsg69hNFRvxJrlUd4mhwRgsrWOgvtJ2BRFyjedK2zhgzh9vSp9XuK+cQR/exj4sjyYYj0K4c1RVCbvDH28N0c1MZRRPgA=="
+      ],
+      "x5t": "rrUKTChB09FDkvkhEmu2Od2r6Ik",
+      "x5t#S256": "lZz-OTZyMutao_jq_6bACLvWzQWApKGDrDrXsLG-cfs"
+    }
+  ]
+}
